Refactoring, service layer improving, fixing the APIs endpoints

+++ NewsBlog/src/main/java/bg/tuvarna/sit/newsblog/service/implement/CommentServiceImpl.java
@@ -1,19 +1,24 @@
 package bg.tuvarna.sit.newsblog.service.implement;
 
-import bg.tuvarna.sit.newsblog.dto.CommentUpdateCreationDto;
+import bg.tuvarna.sit.newsblog.dto.comment.CommentRequestDto;
+import bg.tuvarna.sit.newsblog.dto.comment.CommentResponseDto;
 import bg.tuvarna.sit.newsblog.entity.Comment;
 import bg.tuvarna.sit.newsblog.entity.News;
 import bg.tuvarna.sit.newsblog.entity.User;
 import bg.tuvarna.sit.newsblog.exception.ResourceNotFoundException;
+import bg.tuvarna.sit.newsblog.mapper.CategoryMapper;
+import bg.tuvarna.sit.newsblog.mapper.CommentMapper;
 import bg.tuvarna.sit.newsblog.repository.CommentRepository;
 import bg.tuvarna.sit.newsblog.repository.NewsRepository;
 import bg.tuvarna.sit.newsblog.repository.UserRepository;
 import bg.tuvarna.sit.newsblog.service.interfaces.CommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -21,52 +26,57 @@ public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
     private final NewsRepository newsRepository;
     private final UserRepository userRepository;
+    private final CommentMapper commentMapper;
 
     @Override
-    public Comment createComment(CommentUpdateCreationDto dto, String username, Long newsId) {
-        User user = userRepository.findByUsername(username)
-                //.orElseThrow(() -> new RuntimeException("User not found"));
-                .orElseThrow(()-> new ResourceNotFoundException("User", "name "+username));
-
-        News news = newsRepository.findById(newsId)
-                .orElseThrow(()-> new ResourceNotFoundException("News", "id "+newsId));
-                //.orElseThrow(() -> new RuntimeException("News not found"));
-
-        Comment comment = Comment.builder()
-                .title(dto.getTitle())
-                .content(dto.getContent())
-                .author(user)
-                .news(news)
-                .createdAt(LocalDateTime.now())
-                .build();
-
-        return commentRepository.save(comment);
+    public List<CommentResponseDto> getAllComments() {
+        return commentRepository.findAll().stream()
+                .map(comment -> commentMapper.toDto(comment))
+                .collect(Collectors.toList());
     }
 
     @Override
-    public Comment updateComment(Long id, CommentUpdateCreationDto dto, String username) {
+    public List<CommentResponseDto> getCommentsByNews(Long newsId) {
+        return commentRepository.findByNewsId(newsId).stream()
+                .map(comment -> commentMapper.toDto(comment))
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public CommentResponseDto createComment(Long newsId, CommentRequestDto dto, UserDetails userDetails) {
+        News news = newsRepository.findById(newsId)
+                .orElseThrow(()-> new ResourceNotFoundException("News", "id "+newsId));
+
+        User user = userRepository.findByUsername(userDetails.getUsername())
+                .orElseThrow(()-> new ResourceNotFoundException("User", "name "+userDetails.getUsername()));
+
+        Comment comment = commentMapper.toEntity(dto);
+        commentMapper.enrichComment(comment, news, user);
+
+        return commentMapper.toDto(commentRepository.save(comment));
+    }
+
+    @Override
+    public CommentResponseDto updateComment(Long id, CommentRequestDto dto, UserDetails userDetails) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Comment", "id "+id));
-                //.orElseThrow(() -> new RuntimeException("Comment not found"));
 
-        if (!comment.getAuthor().getUsername().equals(username)) {
+        if (!comment.getAuthor().getUsername().equals(userDetails.getUsername())) {
             throw new RuntimeException("You are not allowed to edit this comment");
         }
 
         comment.setTitle(dto.getTitle());
         comment.setContent(dto.getContent());
 
-        return commentRepository.save(comment);
+        return commentMapper.toDto(commentRepository.save(comment));
     }
 
     @Override
     public void deleteComment(Long id) {
+        if (!commentRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Comment", "id "+id);
+        }
         commentRepository.deleteById(id);
-    }
-
-    @Override
-    public List<Comment> getCommentsByNewsId(Long newsId) {
-        return commentRepository.findByNewsId(newsId);
     }
 }
